Title: 달에서 무게 구하기, Time: 103ms, Memory: 9MB, Status: Passed - CodeTree

+++ 241227/달에서 무게 구하기/weight-on-the-moon.java
@@ -2,7 +2,7 @@ public class Main {
     public static void main(String[] args) {
         // 여기에 코드를 작성해주세요.
         int chew = 13;
-        float ratio = 0.165;
+        double ratio = 0.165;
         System.out.printf("%d * %.6f = %.6f", chew, ratio, chew*ratio);
     }
 }
